Refactoring and upgrade libraries 1. Split into multiple files 2. Upgrade library versions 3. remove rustc-serialize
