+++ .gitignore
@@ -16,7 +16,6 @@
 
 # Generated by Cargo
 /target/
-Cargo.lock
 
 #####=== Vim ===#####
 [._]*.s[a-w][a-z]
