+++ Cargo.toml
@@ -6,16 +6,16 @@ homepage = "https://github.com/sseemayer/keepass-rs"
 repository = "https://github.com/sseemayer/keepass-rs"
 documentation = "https://sseemayer.github.io/keepass-rs/"
 
-version = "0.1.2"
+version = "0.2.0"
 authors = ["Stefan Seemayer <stefan@seemayer.de>"]
 license = "MIT"
 
 readme = "README.md"
 
 [dependencies]
-byteorder = "^0.4.2"
-flate2 = "^0.2.13"
-rust-crypto = "^0.2.34"
-secstr = "^0.2.1"
-xml-rs = "^0.3.0"
-rustc-serialize = "^0.3.16"
+byteorder = "^1"
+flate2 = "^0.2"
+rust-crypto = "^0.2"
+secstr = "^0.3"
+xml-rs = "^0.6"
+base64 = "^0.6"
