+++ src/decrypt.rs
@@ -1,16 +1,16 @@
 use crypto::aes::{cbc_decryptor, ecb_encryptor, KeySize};
 use crypto::blockmodes::{NoPadding, PkcsPadding};
-use crypto::buffer::{RefReadBuffer, RefWriteBuffer, BufferResult, ReadBuffer, WriteBuffer};
+use crypto::buffer::{BufferResult, ReadBuffer, RefReadBuffer, RefWriteBuffer, WriteBuffer};
 use crypto::digest::Digest;
 use crypto::salsa20::Salsa20;
 use crypto::sha2::Sha256;
 use crypto::symmetriccipher::{Decryptor, SymmetricCipherError};
 
-pub trait Cipher {
+pub(crate) trait Cipher {
     fn new(&self, key: &[u8], iv: &[u8]) -> Box<Decryptor>;
 }
 
-pub struct AES256Cipher;
+pub(crate) struct AES256Cipher;
 
 impl Cipher for AES256Cipher {
     fn new(&self, key: &[u8], iv: &[u8]) -> Box<Decryptor> {
@@ -18,7 +18,7 @@ impl Cipher for AES256Cipher {
     }
 }
 
-pub struct Salsa20Cipher;
+pub(crate) struct Salsa20Cipher;
 
 impl Cipher for Salsa20Cipher {
     fn new(&self, key: &[u8], iv: &[u8]) -> Box<Decryptor> {
@@ -26,7 +26,7 @@ impl Cipher for Salsa20Cipher {
     }
 }
 
-pub struct PlainCipher;
+pub(crate) struct PlainCipher;
 
 impl Cipher for PlainCipher {
     fn new(&self, _: &[u8], _: &[u8]) -> Box<Decryptor> {
@@ -37,18 +37,18 @@ impl Cipher for PlainCipher {
 struct NoOpDecryptor;
 
 impl Decryptor for NoOpDecryptor {
-    fn decrypt(&mut self,
-               input: &mut RefReadBuffer,
-               output: &mut RefWriteBuffer,
-               _: bool)
-               -> Result<BufferResult, SymmetricCipherError> {
+    fn decrypt(
+        &mut self,
+        input: &mut RefReadBuffer,
+        output: &mut RefWriteBuffer,
+        _: bool,
+    ) -> Result<BufferResult, SymmetricCipherError> {
         input.push_to(output);
         Ok(BufferResult::BufferUnderflow)
     }
 }
 
-pub fn calculate_sha256(elements: &[&[u8]]) -> [u8; 32] {
-
+pub(crate) fn calculate_sha256(elements: &[&[u8]]) -> [u8; 32] {
     let mut digest = Sha256::new();
 
     for element in elements {
@@ -60,7 +60,7 @@ pub fn calculate_sha256(elements: &[&[u8]]) -> [u8; 32] {
     hash
 }
 
-pub fn decrypt(decryptor: &mut Decryptor, data: &[u8]) -> Result<Vec<u8>, SymmetricCipherError> {
+pub(crate) fn decrypt(decryptor: &mut Decryptor, data: &[u8]) -> Result<Vec<u8>, SymmetricCipherError> {
     let mut final_result = Vec::new();
     let mut read_buffer = RefReadBuffer::new(data);
     let mut buffer = [0u8; 4096];
@@ -68,7 +68,13 @@ pub fn decrypt(decryptor: &mut Decryptor, data: &[u8]) -> Result<Vec<u8>, Symmet
 
     loop {
         let result = try!(decryptor.decrypt(&mut read_buffer, &mut write_buffer, true));
-        final_result.extend(write_buffer.take_read_buffer().take_remaining().iter().cloned());
+        final_result.extend(
+            write_buffer
+                .take_read_buffer()
+                .take_remaining()
+                .iter()
+                .cloned(),
+        );
 
         match result {
             BufferResult::BufferUnderflow => break,
@@ -79,20 +85,19 @@ pub fn decrypt(decryptor: &mut Decryptor, data: &[u8]) -> Result<Vec<u8>, Symmet
     Ok(final_result)
 }
 
-pub fn derive_composite_key(elements: &[&[u8]]) -> [u8; 32] {
+pub(crate) fn derive_composite_key(elements: &[&[u8]]) -> [u8; 32] {
     let credentials_hash = calculate_sha256(elements);
     calculate_sha256(&[&credentials_hash])
 }
 
-pub fn derive_transformed_key(transform_seed: &[u8],
-                              transform_rounds: u64,
-                              composite_key: [u8; 32])
-                              -> Result<[u8; 32], SymmetricCipherError> {
-
+pub(crate) fn derive_transformed_key(
+    transform_seed: &[u8],
+    transform_rounds: u64,
+    composite_key: [u8; 32],
+) -> Result<[u8; 32], SymmetricCipherError> {
     let mut key: [u8; 32] = composite_key.clone();
 
     for _ in 0..transform_rounds {
-
         let mut buffer = [0u8; 32];
         let mut write_buffer = RefWriteBuffer::new(&mut buffer);
 
@@ -102,10 +107,14 @@ pub fn derive_transformed_key(transform_seed: &[u8],
             try!(encryptor.encrypt(&mut read_buffer, &mut write_buffer, true));
         }
 
-        for (&x, k) in write_buffer.take_read_buffer().take_remaining().iter().zip(key.iter_mut()) {
+        for (&x, k) in write_buffer
+            .take_read_buffer()
+            .take_remaining()
+            .iter()
+            .zip(key.iter_mut())
+        {
             *k = x;
         }
-
     }
 
     Ok(calculate_sha256(&[&key]))
