+++ src/xml_parse.rs
@@ -1,18 +1,15 @@
-extern crate crypto;
-extern crate secstr;
-extern crate xml;
 
 use crypto::symmetriccipher::Decryptor;
 
-use rustc_serialize::base64::FromBase64;
+use base64;
 use secstr::SecStr;
 
-use self::xml::reader::{EventReader, XmlEvent};
-use self::xml::name::OwnedName;
+use xml::reader::{EventReader, XmlEvent};
+use xml::name::OwnedName;
 
 use std::collections::HashMap;
 
-use super::{Group, Entry, Value};
+use super::{Entry, Group, Value};
 
 
 enum Node {
@@ -40,9 +37,9 @@ pub fn parse_xml_block(xml: &[u8], decryptor: &mut Decryptor) -> Group {
         match e.unwrap() {
             XmlEvent::StartElement {
                 name: OwnedName { ref local_name, .. },
-                ref attributes, ..
+                ref attributes,
+                ..
             } => {
-
                 xml_stack.push(local_name.clone());
 
                 match &local_name[..] {
@@ -54,60 +51,66 @@ pub fn parse_xml_block(xml: &[u8], decryptor: &mut Decryptor) -> Group {
                         }));
                     }
                     "Entry" => {
-                        parsed_stack.push(Node::Entry(Entry { fields: HashMap::new() }));
+                        parsed_stack.push(Node::Entry(Entry {
+                            fields: HashMap::new(),
+                        }));
                     }
                     "String" => {
-                        parsed_stack.push(Node::KeyValue(String::new(),
-                                                         Value::Unprotected(String::new())));
+                        parsed_stack.push(Node::KeyValue(
+                            String::new(),
+                            Value::Unprotected(String::new()),
+                        ));
                     }
                     "Value" => {
-
                         // Are we encountering a protected value?
-                        if attributes.into_iter()
-                                     .filter(|oa| oa.name.local_name == "Protected")
-                                     .next()
-                                     .map(|oa| &oa.value)
-                                     .map_or(false, |v| {
-                                         v.to_lowercase().parse::<bool>().unwrap_or(false)
-                                     }) {
-
+                        if attributes
+                            .into_iter()
+                            .filter(|oa| oa.name.local_name == "Protected")
+                            .next()
+                            .map(|oa| &oa.value)
+                            .map_or(false, |v| v.to_lowercase().parse::<bool>().unwrap_or(false))
+                        {
                             // Transform value to a Value::Protected
                             if let Some(&mut Node::KeyValue(_, ref mut ev)) =
-                                   parsed_stack.last_mut() {
+                                parsed_stack.last_mut()
+                            {
                                 *ev = Value::Protected(SecStr::new(vec![]));
                             }
                         }
-
                     }
                     _ => {}
                 }
-
             }
 
-            XmlEvent::EndElement { name: OwnedName { ref local_name, .. } } => {
-
+            XmlEvent::EndElement {
+                name: OwnedName { ref local_name, .. },
+            } => {
                 xml_stack.pop();
 
                 if ["Group", "Entry", "String"].contains(&&local_name[..]) {
-
                     let finished_node = parsed_stack.pop().unwrap();
-                    let mut parsed_stack_head = parsed_stack.last_mut();
+                    let parsed_stack_head = parsed_stack.last_mut();
 
                     match finished_node {
                         Node::KeyValue(k, v) => {
-                            if let Some(&mut Node::Entry(Entry {ref mut fields})) =
-                                   parsed_stack_head {
+                            if let Some(&mut Node::Entry(Entry { ref mut fields })) =
+                                parsed_stack_head
+                            {
                                 // A KeyValue was finished inside of an Entry -> add a field
                                 fields.insert(k, v);
                             }
                         }
 
                         Node::Group(finished_group) => {
-                            if let Some(&mut Node::Group(Group {ref mut child_groups, ..})) =
-                                   parsed_stack_head {
+                            if let Some(
+                                &mut Node::Group(Group {
+                                    ref mut child_groups,
+                                    ..
+                                }),
+                            ) = parsed_stack_head
+                            {
                                 // A Group was finished - add Group to parent Group's child groups
                                 child_groups.push(finished_group);
-
                             } else if let None = parsed_stack_head {
                                 // There is no more parent nodes left -> we are at the root
                                 root_group = finished_group;
@@ -115,26 +118,25 @@ pub fn parse_xml_block(xml: &[u8], decryptor: &mut Decryptor) -> Group {
                         }
 
                         Node::Entry(finished_entry) => {
-                            if let Some(&mut Node::Group(Group {ref mut entries, ..})) =
-                                   parsed_stack_head {
+                            if let Some(
+                                &mut Node::Group(Group {
+                                    ref mut entries, ..
+                                }),
+                            ) = parsed_stack_head
+                            {
                                 // A Entry was finished - add Node to parent Group's entries
                                 entries.push(finished_entry);
                             }
                         }
                     }
-
                 }
-
-
             }
 
             XmlEvent::Characters(c) => {
-
                 // Got some character data that need to be matched to a Node on the parsed_stack.
 
                 match (xml_stack.last().map(|s| &s[..]), parsed_stack.last_mut()) {
-                    (Some("Name"),
-                     Some(&mut Node::Group(Group {ref mut name, ..}))) => {
+                    (Some("Name"), Some(&mut Node::Group(Group { ref mut name, .. }))) => {
                         // Got a "Name" element with a Node::Group on the parsed_stack
                         // Update the Group's name
                         *name = c;
@@ -153,11 +155,10 @@ pub fn parse_xml_block(xml: &[u8], decryptor: &mut Decryptor) -> Group {
                                 *v = c;
                             }
                             Value::Protected(ref mut v) => {
-
                                 // Use the decryptor to decrypt the protected
                                 // and base64-encoded value
-                                let buf = &c.from_base64().unwrap()[..];
-                                let buf_decode = super::decrypt::decrypt(decryptor, buf).unwrap();
+                                let buf = base64::decode(&c).unwrap();
+                                let buf_decode = super::decrypt::decrypt(decryptor, buf.as_ref()).unwrap();
                                 let c_decode = String::from_utf8(buf_decode).unwrap();
 
                                 *v = SecStr::from(c_decode);
@@ -169,7 +170,6 @@ pub fn parse_xml_block(xml: &[u8], decryptor: &mut Decryptor) -> Group {
             }
 
             _ => {}
-
         }
     }
